Filmueb version 0.4
ApplicationDatabase class was created to store the records about movies, tv series and actors. ApplicationController class was created to control the flow of the program. Now app has menu which gives an better options for the user. User can add more records, print all of them and exit from the app.